Add implementation switching between Node.js and Elixir
This commit adds the ability to switch between the Node.js and Elixir
implementations using a shared wrapper script. By default, the Node.js
implementation will be used, but the --elixir flag can be provided to
use the Elixir implementation instead.

Implementation details:
- Added wrapper script to select between implementations
- Modified both implementations to display which one is being used
- Updated installation scripts to support both implementations
- Updated documentation with information about implementation switching
- Added --elixir flag to CLI options

Both implementations now display a message on startup to indicate which
implementation is being used.